Moved multitype tests to sparse/tests/test_multitype.py
